♻️ refactor: split pool.go into focused files by responsibility
Extract the 510-line Pool god object into 4 files:
- pool.go (~90 lines): struct, constructor, admin methods (SetFactory, Close)
- pool_session.go (~170 lines): session lifecycle (Create, Resolve, Rotate, Archive, Get, List, History)
- pool_chat.go (~155 lines): Chat streaming orchestration + streamEvents helper
- pool_runner.go (~90 lines): getOrCreateRunner (runner lifecycle, liveness, model switching)

No type or API changes — pure file split. All 77 existing tests pass.
